requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
python-bale-bot==2.5.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==69.0.3
soupsieve==2.5
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
